move deployment history to a folder
