+++ plugins/backstage-plugin-env0/src/components/env0-deployment-table/env0-deployment-table.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import { useGetDeployments } from '../hooks/use-deployments-history';
+import { useGetDeployments } from '../../hooks/use-deployments-history';
 import { Table, TableColumn } from '@backstage/core-components';
-import { Deployment } from '../api/types';
+import { Deployment } from '../../api/types';
 import { PlanSummary } from './plan-summary';
 import { DeploymentType } from './deployment-type';
-import { formatDatetime, parseTimerElapsedTime } from './common/time.utils';
-import dayjs from './common/dayjs.types';
+import { formatDatetime, parseTimerElapsedTime } from '../common/time.utils';
+import dayjs from '../common/dayjs.types';
 import { DeploymentTableHeader } from './deployment-table-header';
-import { ErrorContainer } from './common/error-container';
+import { ErrorContainer } from '../common/error-container';
 
 const columnHeaderStyle = {
   color: '#3636D8',
