+++ plugins/backstage-plugin-env0/src/components/env0-tab-component.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Content, Page } from '@backstage/core-components';
-import { Env0DeploymentTable } from './env0-deployment-table';
+import { Env0DeploymentTable } from './env0-deployment-table/env0-deployment-table';
 import { useEntity } from '@backstage/plugin-catalog-react';
 import { ENV0_ENVIRONMENT_ANNOTATION } from './common/is-plugin-available';
 import { ErrorContainer } from './common/error-container';
